requirements.txt
git+https://github.com/GeoNode/geonode.git@2.8.x#egg=geonode
django-ckeditor==5.3.1
django-bootstrap-breadcrumbs==0.9.1
django-rest-swagger==2.2.0
django-extensions
pandas==0.24.2
django-pandas==0.6.0
django-revproxy==0.9.15
git+https://github.com/voronind/awesome-slugify.git
oauth2-provider==0.0
netCDF4==1.5.1.2
wget==3.2
numpy==1.16.4
fcm-django==0.2.21
#pydap==3.2.2
#sea storm atlas
django-daterange-filter
django-searchable-select
django-grappelli==2.7.3
Django==1.8.19
six==1.10.0
django-cuser==2017.3.16
django-model-utils==3.1.1
django-autocomplete-light==2.3.3
pyshp==1.2.12
Shapely>=1.5.13,<1.6.dev0
OWSLib==0.15.0
proj==0.1.0
pyproj==1.9.5.1
inflection==0.3.1
oauthlib==2.0.1
python-dateutil==2.6.1
pycsw==2.0.3
django-ckeditor==5.3.1
django-bootstrap-breadcrumbs==0.9.1
django-rest-swagger==2.2.0
pygdal==2.1.2.3
pandas==0.24.2
django-pandas==0.6.0
django-revproxy==0.9.15
celery==4.1.0
amqp==2.4.2
kombu==4.1.0
django-celery-beat==1.1.1
django-celery-monitor==1.1.2
django-celery-results==1.0.1
#GeoNode==2.8
